Compress weight delta to float16 and fix MPS sync overhead in gradient compression
leader.py: delta saved as float16 (43MB -> 22MB per step). Worker's add_()
handles mixed precision automatically -- no worker change needed.

worker.py: flush MPS/CUDA command queue once before the per-layer CPU
transfer loop in compress_gradients. Previously each .cpu() call triggered
an implicit device sync; now one explicit sync covers all 60+ layers.

+++ leader.py
@@ -830,10 +830,13 @@ class LeaderService(trainer_pb2_grpc.TrainerServiceServicer):
         if losses:
             self._global_loss = sum(losses) / len(losses)
 
-        # Compute parameter delta (new - old) and serialise
+        # Compute parameter delta (new - old) and serialise as float16
+        # float32 → float16 halves transfer size (43 MB → 22 MB) with negligible
+        # precision loss for typical LR-scaled deltas (magnitude << 1.0).
+        # apply_delta() on the worker uses add_() which handles mixed precision.
         new_state  = self._model.state_dict()
         delta = {
-            k: (v.cpu() - self._prev_state[k])
+            k: (v.cpu() - self._prev_state[k]).half()
             for k, v in new_state.items()
             if k in self._prev_state
         }
